refactor(client): extract the history component from the chatroom component
